+++ cmn-data-svc-server/db-scripts/cds-mysql-downgrade-1.15-to-1.14.sql
@@ -20,6 +20,15 @@
 -- FROM version 1.15.x TO version 1.14.x.
 -- No database is specified to allow flexible deployment!
 
+-- 10
+ALTER TABLE C_SOLUTION_GROUP
+  DROP INDEX C_SOLUTION_GROUP_C_NAME;
+-- 9
+ALTER TABLE C_PEER_GROUP
+  DROP INDEX C_PEER_GROUP_C_NAME;
+-- 8
+ALTER TABLE C_ROLE
+  DROP INDEX C_ROLE_C_NAME;
 -- 7
 UPDATE C_SITE_CONFIG set CONFIG_VAL = 
   '{"fields":[ {"type":"text","name":"siteInstanceName","label":"siteInstanceName","required":"true","data":"Acumos"}, {"type":"file","name":"headerLogo","label":"Headerlogo","data":{"lastModified":1510831880727,"lastModifiedDate":"2017-11-16T11:31:20.727Z","name":"acumos_logo_white.png","size":3657,"type":"image/png"}}, {"type":"file","name":"footerLogo","label":"Footerlogo","data":{"lastModified":1510831874776,"lastModifiedDate":"2017-11-16T11:31:14.776Z","name":"footer_logo.png","size":3127,"type":"image/png"}}, {"type":"heading","name":"ConnectionConfig","label":"ConnectionConfig","required":"true","subFields":[ {"type":"text","name":"socketTimeout","label":"socketTimeout","required":"true","data":"300"}, {"type":"text","name":"connectionTimeout","label":"connectionTimeout","required":"true","data":"10"}]}, {"type":"select","name":"enableOnBoarding","label":"EnableOnboarding","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}]}'