+++ cmn-data-svc-server/db-scripts/cds-mysql-upgrade-1.14-to-1.15.sql
@@ -18,8 +18,16 @@
 
 -- Script to upgrade database used by the Common Data Service
 -- FROM version 1.14.x TO version 1.15.x.
--- No database is specified to allow flexible deployment!
+-- No database name is set to allow flexible deployment.
 
+-- WARNING!!  THIS SCRIPT ADDS UNIQUE CONSTRAINTS TO C_ROLE, C_PEER_GROUP and C_SOLUTION_GROUP. 
+-- IF ANY OF THE FOLLOWING SELECTS RETURNS A RECORD THEN THE UNIQUE CONSTRAINT CREATION WILL FAIL, 
+-- FAILING THE SCRIPT. IN THAT CASE MODIFY THE RECORDS IN THE DATABASE WITH DUPLICATE ROLE NAME
+-- AND/OR PEER GROUP NAME AND/OR SOLUTION GROUP NAME UNTIL THE NAMES ARE UNIQUE, THEN RERUN.
+-- 0
+SELECT NAME, count(*) AS c FROM C_ROLE           GROUP BY NAME HAVING c > 1;
+SELECT NAME, count(*) AS c FROM C_PEER_GROUP     GROUP BY NAME HAVING c > 1;
+SELECT NAME, count(*) AS c FROM C_SOLUTION_GROUP GROUP BY NAME HAVING c > 1;
 -- 1
 ALTER TABLE C_SOLUTION_REV
   ADD COLUMN ACCESS_TYPE_CD CHAR(2) NOT NULL DEFAULT 'PR';
@@ -54,3 +62,9 @@ ALTER TABLE C_SOLUTION
 UPDATE C_SITE_CONFIG set CONFIG_VAL = 
   '{"fields":[ {"type":"text","name":"siteInstanceName","label":"siteInstanceName","required":"true","data":"Acumos"}, {"type":"file","name":"headerLogo","label":"Headerlogo","data":{"lastModified":1510831880727,"lastModifiedDate":"2017-11-16T11:31:20.727Z","name":"acumos_logo_white.png","size":3657,"type":"image/png"}}, {"type":"file","name":"footerLogo","label":"Footerlogo","data":{"lastModified":1510831874776,"lastModifiedDate":"2017-11-16T11:31:14.776Z","name":"footer_logo.png","size":3127,"type":"image/png"}}, {"type":"heading","name":"ConnectionConfig","label":"ConnectionConfig","required":"true","subFields":[ {"type":"text","name":"socketTimeout","label":"socketTimeout","required":"true","data":"300"}, {"type":"text","name":"connectionTimeout","label":"connectionTimeout","required":"true","data":"10"} ]}, {"type":"select","name":"enableOnBoarding","label":"EnableOnboarding","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}, {"type":"textarea","name":"validationText","label":"validationText","required":"false","data":"test"}, {"type":"select","name":"EnableDCAE","label":"EnableDCAE","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}]}'
   WHERE CONFIG_KEY='site_config';
+-- 8
+CREATE UNIQUE INDEX C_ROLE_C_NAME ON C_ROLE (NAME);
+-- 9
+CREATE UNIQUE INDEX C_PEER_GROUP_C_NAME ON C_PEER_GROUP (NAME);
+-- 10
+CREATE UNIQUE INDEX C_SOLUTION_GROUP_C_NAME ON C_SOLUTION_GROUP (NAME);