+++ cmn-data-svc-server/db-scripts/cmn-data-svc-ddl-dml-mysql-1.15.sql
@@ -28,8 +28,8 @@ CREATE TABLE C_USER (
   MIDDLE_NAME VARCHAR(50),
   LAST_NAME VARCHAR(50),
   ORG_NAME VARCHAR(50),
-  EMAIL VARCHAR(100) UNIQUE,
-  LOGIN_NAME VARCHAR(25) NOT NULL UNIQUE,
+  EMAIL VARCHAR(100),
+  LOGIN_NAME VARCHAR(25) NOT NULL,
   LOGIN_HASH VARCHAR(64),
   LOGIN_PASS_EXPIRE_DATE DATETIME NULL default NULL,
   -- JSON web token
@@ -38,7 +38,9 @@ CREATE TABLE C_USER (
   LAST_LOGIN_DATE DATETIME NULL DEFAULT NULL,
   PICTURE BLOB,
   CREATED_DATE TIMESTAMP NOT NULL DEFAULT 0,
-  MODIFIED_DATE TIMESTAMP NOT NULL
+  MODIFIED_DATE TIMESTAMP NOT NULL,
+  UNIQUE INDEX C_USER_C_EMAIL (EMAIL),
+  UNIQUE INDEX C_USER_C_LOGIN_NAME (LOGIN_NAME)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE C_ROLE (
@@ -46,7 +48,8 @@ CREATE TABLE C_ROLE (
   NAME VARCHAR(100) NOT NULL,
   ACTIVE_YN CHAR(1) DEFAULT 'Y' NOT NULL,
   CREATED_DATE TIMESTAMP NOT NULL DEFAULT 0,
-  MODIFIED_DATE TIMESTAMP NOT NULL
+  MODIFIED_DATE TIMESTAMP NOT NULL,
+  UNIQUE INDEX C_ROLE_C_NAME (NAME)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 -- Many:many mapping of user to role requires a map (join) table
@@ -387,7 +390,8 @@ CREATE TABLE C_PEER_GROUP (
   NAME VARCHAR(50) NOT NULL,
   DESCRIPTION VARCHAR(512),
   CREATED_DATE TIMESTAMP NOT NULL DEFAULT 0,
-  MODIFIED_DATE TIMESTAMP NOT NULL
+  MODIFIED_DATE TIMESTAMP NOT NULL,
+  UNIQUE INDEX C_PEER_GROUP_C_NAME (NAME)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE C_SOLUTION_GROUP (
@@ -395,7 +399,8 @@ CREATE TABLE C_SOLUTION_GROUP (
   NAME VARCHAR(50) NOT NULL,
   DESCRIPTION VARCHAR(512),
   CREATED_DATE TIMESTAMP NOT NULL DEFAULT 0,
-  MODIFIED_DATE TIMESTAMP NOT NULL
+  MODIFIED_DATE TIMESTAMP NOT NULL,
+  UNIQUE INDEX C_SOLUTION_GROUP_C_NAME (NAME)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE C_PEER_GRP_MEM_MAP (
@@ -444,7 +449,7 @@ INSERT INTO C_ROLE (ROLE_ID, NAME, ACTIVE_YN) VALUES ('12345678-abcd-90ab-cdef-1
 # Unfortunately JSON does not allow embedded newlines.
 INSERT INTO C_SITE_CONFIG (CONFIG_KEY, CONFIG_VAL) VALUES (
   'site_config',
-  '{"fields":[ {"type":"text","name":"siteInstanceName","label":"siteInstanceName","required":"true","data":"Acumos"}, {"type":"file","name":"headerLogo","label":"Headerlogo","data":{"lastModified":1510831880727,"lastModifiedDate":"2017-11-16T11:31:20.727Z","name":"acumos_logo_white.png","size":3657,"type":"image/png"}}, {"type":"file","name":"footerLogo","label":"Footerlogo","data":{"lastModified":1510831874776,"lastModifiedDate":"2017-11-16T11:31:14.776Z","name":"footer_logo.png","size":3127,"type":"image/png"}}, {"type":"heading","name":"ConnectionConfig","label":"ConnectionConfig","required":"true","subFields":[ {"type":"text","name":"socketTimeout","label":"socketTimeout","required":"true","data":"300"}, {"type":"text","name":"connectionTimeout","label":"connectionTimeout","required":"true","data":"10"} ]}, {"type":"select","name":"enableOnBoarding","label":"EnableOnboarding","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}, {"type":"textarea","name":"validationText","label":"validationText","required":"false","data":"test"}, {"type":"select","name":"EnableDCAE","label":"EnableDCAE","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}]}'
+  '{"fields":[{"type":"text","name":"siteInstanceName","label":"siteInstanceName","required":"true","data":"Acumos"}, {"type":"file","name":"headerLogo","label":"Headerlogo","data":{"lastModified":1510831880727,"lastModifiedDate":"2017-11-16T11:31:20.727Z","name":"acumos_logo_white.png","size":3657,"type":"image/png"}},{"type":"file","name":"footerLogo","label":"Footerlogo","data":{"lastModified":1510831874776,"lastModifiedDate":"2017-11-16T11:31:14.776Z","name":"footer_logo.png","size":3127,"type":"image/png"}},{"type":"heading","name":"ConnectionConfig","label":"ConnectionConfig","required":"true","subFields":[{"type":"text","name":"socketTimeout","label":"socketTimeout","required":"true","data":"300"},{"type":"text","name":"connectionTimeout","label":"connectionTimeout","required":"true","data":"10"}]},{"type":"select","name":"enableOnBoarding","label":"EnableOnboarding","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}},{"type":"textarea","name":"validationText","label":"validationText","required":"false","data":"test"},{"type":"select","name":"EnableDCAE","label":"EnableDCAE","options":[{"name":"Enabled"},{"name":"Disabled"}],"required":true,"data":{"name":"Enabled"}}]}'
 );
 
 # Initial user, role setups for admin